requirements.txt
pypiwin32==223
attrs~=19.1.0
selenium~=3.141.0
openpyxl~=2.6.3
ddt~=1.2.1
PyMySQL~=0.9.3
pytz~=2019.2
pyyaml~=5.3.1
pytest~=4.0.2
pytest-html~=2.1.1
pytest-repeat~=0.8.0
pytest-rerunfailures~=9.0
pytest-assume~=2.2.1
pytest-xdist~=1.32.0
pytest-ordering~=0.6
allure-pytest~=2.8.18
python-dateutil~=2.8.0
jsonpath~=0.82
requests~=2.25.1